refactor: 오목을 판별하는 로직을 FiveRule 클래스로 분리
